packaging processing into classes, renaming
