+++ docker/depth_stage/entrypoint.sh
