+++ docker/embeddings_stage/entrypoint.sh
