+++ docker/filter_stage/entrypoint.sh
