+++ docker/prompt_stage/Dockerfile
@@ -1,8 +1,8 @@
 FROM smellslikeml/vqasynth:base
 WORKDIR /app
 
-COPY docker/pointcloud_processor/process_pointcloud.py /app
-COPY docker/pointcloud_processor/entrypoint.sh /app
+COPY docker/prompt_stage/process_prompts.py /app
+COPY docker/prompt_stage/entrypoint.sh /app
 COPY config/config.yaml /app/config/config.yaml
 
 RUN pip install pandas
