+++ vqasynth/datasets/depth.py
@@ -1,8 +1,8 @@
 import cv2
 import numpy as np
 from PIL import Image
+
 import depth_pro
-from vqasynth.datasets.utils import colorize
 
 class DepthEstimator:
     def __init__(self):
@@ -10,9 +10,9 @@ class DepthEstimator:
         self.model, self.transform = depth_pro.create_model_and_transforms()
         self.model.eval()
 
-    def run_inference(self, image_path):
+    def run(self, image_path):
         """
-        Takes a Pillow RGB image, preprocesses it, and returns the depth and focal length in pixels.
+        Returns a depth map and formatted focal length from an image
 
         Args:
             image_path: path to image
@@ -20,11 +20,20 @@ class DepthEstimator:
         Returns:
             tuple: depth in meters, focal length in pixels
         """
-        image, _, f_px = depth_pro.load_rgb(image_path)
-        image_tensor = self.transform(image)
-        prediction = self.model.infer(image_tensor, f_px=f_px)
-        depth = prediction["depth"]
-        depth_normalized = cv2.normalize(np.array(depth), None, 255, 0, cv2.NORM_MINMAX, cv2.CV_8U)
-        focallength_px = prediction["focallength_px"] 
 
-        return depth_normalized, focallength_px.item()
+        try:
+            image, _, f_px = depth_pro.load_rgb(image_path)
+            image_tensor = self.transform(image)
+            prediction = self.model.infer(image_tensor, f_px=f_px)
+            depth = prediction["depth"]
+            depth_map = cv2.normalize(np.array(depth), None, 255, 0, cv2.NORM_MINMAX, cv2.CV_8U)
+            focallength_px = prediction["focallength_px"] 
+
+            if depth_map.dtype != np.uint16:
+                depth_map = (depth_map / np.max(depth_map) * 65535).astype(np.uint16)
+                depth_image = Image.fromarray(depth_map, mode='I;16')
+
+            return depth_image, focallength_px
+        except Exception as e:
+            print(f"Error during segmentation: {str(e)}")
+            return [], 0
